refactor: colonizethis_turn wave-5 slice C lib module splits (#4168)
Split spy_resolver, naval_resolution_move, turn_news_digest, and
world_market_phase_orders into focused sibling modules while preserving
public exports and characterization test behavior.

Refs #4168
